Made all requested changes
